fix: navbar banner layout shift
