+++ src/components/app/geoLocationDisclaimerBanner/geoLocationDisclaimerBanner.tsx
@@ -1,22 +1,17 @@
 'use client'
 
 import { useEffect, useState } from 'react'
-import Balancer from 'react-wrap-balancer'
 import Cookies from 'js-cookie'
-import { ArrowRight } from 'lucide-react'
 import Link from 'next/link'
 import { useRouter } from 'next/navigation'
 
-import { InternalLink } from '@/components/ui/link'
 import { useHasHydrated } from '@/hooks/useHasHydrated'
 import { useIsMobile } from '@/hooks/useIsMobile'
-import { useLocale } from '@/hooks/useLocale'
 import {
   parseUserCountryCodeCookie,
   USER_COUNTRY_CODE_COOKIE_NAME,
 } from '@/utils/server/getCountryCode'
 import { SUPPORTED_COUNTRY_CODES } from '@/utils/shared/supportedCountries'
-import { getIntlUrls } from '@/utils/shared/urls'
 
 const languages = getNavigatorLanguages()
 
@@ -25,8 +20,6 @@ export function GeoLocationDisclaimerBanner() {
   const router = useRouter()
   const isMobile = useIsMobile()
   const [isVisible, setIsVisible] = useState(false)
-  const locale = useLocale()
-  const urls = getIntlUrls(locale)
 
   const WrapperContainer = isMobile ? 'button' : 'div'
 
@@ -92,21 +85,7 @@ export function GeoLocationDisclaimerBanner() {
     )
   }
 
-  return (
-    <InternalLink
-      className="flex w-full items-center bg-primary-cta text-center opacity-100 transition-all duration-200"
-      href={urls.emailDeeplink()}
-    >
-      <div className="container flex w-full justify-between">
-        <div className="flex w-full items-center justify-center py-2 text-center text-base text-white antialiased">
-          <Balancer>
-            Make your voice heard — tell your Senators to vote NO on an anti-crypto SEC nominee!
-          </Balancer>
-          <ArrowRight className="w-12 lg:w-8" size={16} />
-        </div>
-      </div>
-    </InternalLink>
-  )
+  return null
 }
 
 /**
